move invalid_message_halt_test to separate package to avoid test pollution
